fix some compile bugs

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: kdaumont <kdaumont@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 15:11:39 by kdaumont          #+#    #+#             */
-/*   Updated: 2023/11/07 09:05:44 by kdaumont         ###   ########.fr       */
+/*   Updated: 2023/11/09 12:58:47 by kdaumont         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*mem;
 
 	i = 0;
